validation

+++ src/main/java/com/demo/ecommerce/model/ItemFactory.java
@@ -2,7 +2,7 @@ package com.demo.ecommerce.model;
 
 public class ItemFactory {
 
-    public static Item create(String id, String name, int qty, double price) {
+    public static Item create(String id, String name, int qty, int price) {
         return new Item( name, qty, price);
     }
 }
